refactor the team section
