+++ src/components/index.tsx
@@ -14,7 +14,7 @@ export { Menu } from './Navbar/Menu';
  *
  */
 export { HeroStatSection } from './sections/HeroStats.section';
-export { TeamSection } from './sections/Team.section';
+export { TeamSection } from './sections/Team/Team.section';
 export { ContactSection } from './sections/Contact.section';
 export { FooterSection } from './sections/Footer.section';
 export { SponsorFAQSection } from './sections/SponsorFAQ.section';
