+++ src/components/sections/Team/data.ts
@@ -1,16 +1,3 @@
-interface Member {
-    name: string;
-    title: string;
-    profile: string;
-}
-
-interface MemberProps extends Member {}
-
-interface DepartmentProps {
-    title: string;
-    members: Member[];
-}
-
 const departments = [
     {
         title: 'design',
@@ -144,91 +131,4 @@ const departments = [
     },
 ];
 
-const Member: React.FC<MemberProps> = ({ name, title, profile }) => {
-    return (
-        // ?width can be fixed (84 when small, 128 when big or large viewport) (done)
-        <li className="flex items-center gap-x-4">
-            <img
-                className="aspect-square w-20 rounded-full object-cover lg:w-32"
-                src={profile}
-                alt={title}
-            />
-
-            <div>
-                <p className="text-lg lg:text-2xl font-extrabold">{name}</p>
-                <p className="text-base lg:text-lg font-normal">{title}</p>
-            </div>
-        </li>
-    );
-};
-
-const Department: React.FC<DepartmentProps> = ({ title, members }) => {
-    return (
-        <div className="department">
-            <h3 className="mb-6 text-3xl font-bold capitalize md:mb-8">
-                {title}
-            </h3>
-            <ul className="grid gap-4 sm:grid-cols-2 md:gap-6 lg:grid-cols-3 lg:gap-8">
-                {members.map((member, index) => (
-                    <Member
-                        {...member}
-                        key={`${index}-${member.name}-${member.title}`}
-                    />
-                ))}
-            </ul>
-        </div>
-    );
-};
-
-const TeamSection = () => {
-    return (
-        // should have decide the consistent padding
-        // or right now, i can focus on the section and increase the padding
-        <section
-            id="team-section"
-            className="relative isolate overflow-hidden bg-[#9f7eca] p-5 md:p-8 lg:p-12"
-        >
-            {/* image at the top */}
-            <div className="relative mx-auto mb-10 sm:mb-24 w-fit lg:mb-32">
-                {/* keep the banner only in mobile view */}
-                {/* keep the balloon after mobile view */}
-                <img
-                    className="mx-auto w-full max-w-fit sm:hidden"
-                    src="src/assets/team/meet-the-team-banner.webp"
-                    alt=""
-                />
-                <img
-                    className="mx-auto hidden w-full max-w-[70rem] sm:block"
-                    src="src/assets/team/meet-the-team-balloon-banner.webp"
-                    alt=""
-                />
-                <img
-                    className="absolute bottom-0 translate-y-1/2  z-10 hidden sm:block"
-                    src="src/assets/team/clouds.svg"
-                    alt=""
-                />
-            </div>
-
-            {/* right building image */}
-            <img
-                className="absolute right-0 -z-10 hidden translate-x-1/3 lg:block 2xl:translate-x-32"
-                src="src/assets/team/meet-the-team-building-right.svg"
-                alt=""
-            />
-            {/* left building image */}
-            <img
-                className="absolute bottom-0 left-0 -z-10 hidden -translate-x-1/3 translate-y-1/2 lg:block 2xl:translate-y-[30%]"
-                src="src/assets/team/meet-the-team-building-left.svg"
-                alt=""
-            />
-
-            <div className="mx-auto max-w-[1328px] space-y-12 pb-80">
-                {departments.map((department) => (
-                    <Department {...department} key={department.title} />
-                ))}
-            </div>
-        </section>
-    );
-};
-
-export { TeamSection };
+export { departments };
